Implementação login

+++ src/controllers/UserController.ts
@@ -1,67 +1,71 @@
-import db from '../database/models';
-//import {UserType} from '../database/models/users';
+import { Request, Response } from "express";
+
+import db from "../database/models";
+import { Users } from "../database/models/users";
 
 class UserController {
+	constructor () { }
 
-	constructor() {}
-
-	public static async getAllUsers(req:any, res:any) {
+	public static async getAllUsers (req: Request, res: Response) {
 		try {
-			const allUsers = await db.Users.findAll();
+			const allUsers = await db.Users.findAll({
+				attributes: ["idUser", "name", "email", "type"]
+			});
 			return res.status(200).json(allUsers);
 		} catch (err) {
 			return res.status(500).json(err.message);
 		}
 	}
 
-	public static async getOneUserById(req:any, res:any) {
+	public static async getOneUserById (req: Request, res: Response) {
 		const { idUser } = req.params;
 		try {
-			const user = await db.Users.findOne( { where: { idUser:Number(idUser) } } );
+			const user = await db.Users.findOne({
+				attributes: ["idUser", "name", "email", "type"],
+				where: { idUser: Number(idUser) }
+			});
 			return res.status(200).json(user);
 		} catch (err) {
 			return res.status(500).json(err.message);
 		}
 	}
 
-	public static async getOneUserByLogin(req:any, res:any) {
-		const { email, password } = req.body;
-		try {
-			const user = await db.Users.findOne( { where: {email:String(email), password:String(password) } } );
-			return res.status(200).json(user);
-		} catch (err) {
-			return res.status(500).json(err.message);
-		}
-	}
-
-	public static async createUser(req:any, res:any) {
+	public static async createUser (req: Request, res: Response) {
 		const newUser = req.body;
-		//let userType = type as UserType;
 		try {
-			const newUserCreated = await db.Users.create( newUser );
+			const newUserCreated = (await db.Users.create(newUser)).toJSON() as Partial<Users>;
+			delete newUserCreated.password;
 			return res.status(200).json(newUserCreated);
 		} catch (err) {
 			return res.status(500).json(err.message);
 		}
 	}
 
-	public static async updateUser(req:any, res:any) {
+	public static async updateUser (req: Request, res: Response) {
 		const { idUser } = req.params;
 		const newInfo = req.body;
 		try {
-			await db.Users.update( newInfo, { where: { idUser:Number(idUser) } } );
-			const updatedUser = await db.Users.findOne( { where: { idUser:Number(idUser) } } );
+			await db.Users.update(newInfo, { where: { idUser: Number(idUser) } });
+			const updatedUser = await db.Users.findOne({
+				attributes: ["idUser", "name", "email", "type"],
+				where: { idUser: Number(idUser) }
+			});
 			return res.status(200).json(updatedUser);
 		} catch (err) {
 			return res.status(500).json(err.message);
 		}
 	}
 
-	public static async deleteUser(req:any, res:any) {
+	public static async deleteUser (req: Request, res: Response) {
 		const { idUser } = req.params;
+
+		const authUser = res.locals.user as Partial<Users>;
+		if (authUser.idUser != Number(idUser))
+			return res.status(403).json({ message: "Você não pode deletar outro usuário, apenas o seu próprio!" });
+
 		try {
-			await db.Users.destroy( { where: { idUser:Number(idUser) } } );
-			return res.status(200).json( {message: `id ${idUser} deleted` } );
+			await db.Users.destroy({ where: { idUser: Number(idUser) } });
+			return res.status(200).json({ message: `Usuário de ID ${idUser} deletado.` });
 		} catch (err) {
 			return res.status(500).json(err.message);
 		}
